requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2.1
django-import-export==4.3.7
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.2.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
